Pass the array element length as a full int, not int16
Reading or writing an element of a non-native array whose element type is
pass-by-reference -- numeric[], text[] -- aborted array_get_element on an
assert-enabled build ("Assert(elmlen == eah->typlen)"), and returned a wrong
value on a non-assert build.  A varlena element type has elmlen = -1, and the
element helpers declared that argument as int16 with no sign-extension
attribute on the LLVM call; on arm64 the negative value arrived at the C helper
zero-extended to 65535, which array_get_element rejects.  Fixed-length elements
(int4, int8, float8) are positive and were unaffected, which is why only
numeric/text arrays were hit.

Widen elmlen to a full int (i32) in uplpgsql_rt_array_get_element,
uplpgsql_rt_array_set_element, their LLVM function types, and the constant
resolve_array_type_info emits -- sidestepping the ABI ambiguity rather than
relying on a sign-extension attribute.

This is a long-standing bug, surfaced by PR #39's out-of-range read test.  With
it fixed, those tests move back to numeric[] and a numeric/text section is
added covering reads (in range, NULL element, out of range, NULL subscript)
and NULL writes.  plpgsql_array runs byte-identical to a stock PL/pgSQL mirror.

+++ common/upl_common.h
@@ -479,14 +479,14 @@ extern void uplpgsql_emit_refresh_native_array(UPLpgSQL_compile_ctx *ctx,
 /* Runtime helpers for array element access from inlined expressions */
 extern Datum uplpgsql_rt_array_get_element(UPLpgSQL_exec_state *estate,
 										   int array_dno, int subscript,
-										   int typlen, int16 elmlen,
+										   int typlen, int elmlen,
 										   bool elmbyval, char elmalign,
 										   bool *isNull_out);
 extern void uplpgsql_rt_array_set_element(UPLpgSQL_exec_state *estate,
 										  int array_dno, int subscript,
 										  Datum newvalue, bool valisnull,
 										  int typlen, Oid elemtype,
-										  int16 elmlen, bool elmbyval,
+										  int elmlen, bool elmbyval,
 										  char elmalign);
 
 /* Runtime helpers for inlined overflow/division errors */
